remove IPersistenceTestsConfiguration
