--- src/NServiceBus.Storage.MongoDB.Tests/ComponentTests/IPersistenceTestsConfiguration.cs
@@ -1,48 +1,0 @@
-﻿// ReSharper disable UnusedMemberInSuper.Global
-
-namespace NServiceBus.Persistence.ComponentTests
-{
-    using System;
-    using System.Threading.Tasks;
-    using Outbox;
-    using Sagas;
-    using Timeout.Core;
-    using Unicast.Subscriptions.MessageDrivenSubscriptions;
-
-    public interface IPersistenceTestsConfiguration
-    {
-        bool SupportsDtc { get; }
-
-        bool SupportsOutbox { get; }
-
-        bool SupportsFinders { get; }
-
-        bool SupportsSubscriptions { get; }
-
-        bool SupportsTimeouts { get; }
-
-        bool SupportsPessimisticConcurrency { get; }
-
-        ISagaIdGenerator SagaIdGenerator { get; }
-
-        ISagaPersister SagaStorage { get; }
-
-        ISynchronizedStorage SynchronizedStorage { get; }
-
-        ISynchronizedStorageAdapter SynchronizedStorageAdapter { get; }
-
-        ISubscriptionStorage SubscriptionStorage { get; }
-
-        IPersistTimeouts TimeoutStorage { get; }
-
-        IQueryTimeouts TimeoutQuery { get; }
-
-        IOutboxStorage OutboxStorage { get; }
-
-        Task Configure();
-
-        Task Cleanup();
-
-        Task CleanupMessagesOlderThan(DateTimeOffset beforeStore);
-    }
-}
