+++ src/NServiceBus.Storage.MongoDB.Tests/ComponentTests/PersistenceTestsConfiguration.cs
@@ -3,12 +3,20 @@
 namespace NServiceBus.Persistence.ComponentTests
 {
     using System;
+    using System.Globalization;
     using System.Linq;
     using System.Reflection;
+    using System.Threading.Tasks;
     using Extensibility;
+    using global::MongoDB.Driver;
+    using Outbox;
     using Sagas;
+    using Storage.MongoDB;
+    using Storage.MongoDB.Tests;
+    using Timeout.Core;
+    using Unicast.Subscriptions.MessageDrivenSubscriptions;
 
-    public partial class PersistenceTestsConfiguration : IPersistenceTestsConfiguration
+    public class PersistenceTestsConfiguration
     {
         public Func<ContextBag> GetContextBagForTimeoutPersister { get; set; } = () => new ContextBag();
         public Func<ContextBag> GetContextBagForSagaStorage { get; set; } = () => new ContextBag();
@@ -31,5 +39,99 @@ namespace NServiceBus.Persistence.ComponentTests
         }
 
         SagaMetadataCollection sagaMetadataCollection;
+
+        public PersistenceTestsConfiguration(string versionElementName, Func<Type, string> collectionNamingConvention, TimeSpan? transactionTimeout = null)
+        {
+            DatabaseName = "Test_" + DateTime.Now.Ticks.ToString(CultureInfo.InvariantCulture);
+            CollectionNamingConvention = collectionNamingConvention;
+
+            SynchronizedStorage = new StorageSessionFactory(ClientProvider.Client, true, DatabaseName, collectionNamingConvention, transactionTimeout.HasValue ? transactionTimeout.Value : MongoPersistence.DefaultTransactionTimeout);
+            SynchronizedStorageAdapter = new StorageSessionAdapter();
+
+            SagaIdGenerator = new DefaultSagaIdGenerator();
+            SagaStorage = new SagaPersister(versionElementName);
+
+            OutboxStorage = new OutboxPersister(ClientProvider.Client, DatabaseName, collectionNamingConvention);
+
+            var subscriptionCollection = ClientProvider.Client.GetDatabase(DatabaseName, MongoPersistence.DefaultDatabaseSettings).GetCollection<EventSubscription>("eventsubscriptions");
+            var subscriptionPersister = new SubscriptionPersister(subscriptionCollection);
+            subscriptionPersister.CreateIndexes();
+            SubscriptionStorage = subscriptionPersister;
+        }
+
+        public PersistenceTestsConfiguration(TimeSpan? transactionTimeout = null) : this("_version", t => t.Name.ToLower(), transactionTimeout)
+        {
+        }
+
+        public PersistenceTestsConfiguration(string versionElementName) : this(versionElementName, t => t.Name.ToLower())
+        {
+        }
+
+        public string DatabaseName { get; }
+
+        public Func<Type, string> CollectionNamingConvention { get; }
+
+        public bool SupportsDtc { get; } = false;
+
+        public bool SupportsOutbox { get; } = true;
+
+        public bool SupportsFinders { get; } = true;
+
+        public bool SupportsSubscriptions { get; } = true;
+
+        public bool SupportsTimeouts { get; } = false;
+
+        public bool SupportsPessimisticConcurrency { get; } = true;
+
+        public ISagaIdGenerator SagaIdGenerator { get; }
+
+        public ISagaPersister SagaStorage { get; }
+
+        public ISynchronizedStorage SynchronizedStorage { get; }
+
+        public ISynchronizedStorageAdapter SynchronizedStorageAdapter { get; }
+
+        public ISubscriptionStorage SubscriptionStorage { get; }
+
+        public IPersistTimeouts TimeoutStorage { get; }
+
+        public IQueryTimeouts TimeoutQuery { get; }
+
+        public IOutboxStorage OutboxStorage { get; }
+
+        public async Task Configure()
+        {
+            var databaseSettings = new MongoDatabaseSettings
+            {
+                ReadConcern = ReadConcern.Majority,
+                ReadPreference = ReadPreference.Primary,
+                WriteConcern = WriteConcern.WMajority
+            };
+
+            var database = ClientProvider.Client.GetDatabase(DatabaseName, databaseSettings);
+
+            await database.CreateCollectionAsync(CollectionNamingConvention(typeof(OutboxRecord)));
+
+            Storage.MongoDB.SagaStorage.InitializeSagaDataTypes(ClientProvider.Client, DatabaseName, CollectionNamingConvention, SagaMetadataCollection);
+        }
+
+        public async Task Cleanup()
+        {
+            await ClientProvider.Client.DropDatabaseAsync(DatabaseName);
+        }
+
+        public Task CleanupMessagesOlderThan(DateTimeOffset beforeStore)
+        {
+            return Task.FromResult(0);
+        }
+
+        class DefaultSagaIdGenerator : ISagaIdGenerator
+        {
+            public Guid Generate(SagaIdGeneratorContext context)
+            {
+                // intentionally ignore the property name and the value.
+                return CombGuid.Generate();
+            }
+        }
     }
 }
